# bar.openapi.yml
openapi: 3.0.3
info:
  title: OpenAPI - Bar
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.11
servers:
  - url: http://localhost:8080
  - url: https://le-bar.telecomnancy.net/api/v1
...
paths:
...
  /transactions:
    get:
      description: Get all active transactions (orders)
      summary: ''
      operationId: getTransactions
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            format: uint64
        - name: limit
          in: query
          description: Number of transactions per page
          required: false
          schema:
            type: integer
            format: uint64
        - name: state
          in: query
          description: Filter by state
          required: false
          schema:
            type: string
            $ref: '#/components/schemas/TransactionState'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  page:
                    type: integer
                    format: uint64
                  limit:
                    type: integer
                    format: uint64
                  max_page:
                    type: integer
                    format: uint64
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
  /account/transactions:
    get:
      summary: ''
      description: Get all transactions
      operationId: getCurrentAccountTransactions
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            format: uint64
        - name: limit
          in: query
          description: Number of transactions per page
          required: false
          schema:
            type: integer
            format: uint64
        - name: state
          in: query
          description: Filter by state
          required: false
          schema:
            type: string
            $ref: '#/components/schemas/TransactionState'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  page:
                    type: integer
                    format: uint64
                  limit:
                    type: integer
                    format: uint64
                  max_page:
                    type: integer
                    format: uint64
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - auth: []
      tags:
        - transactions
    post:
      summary: ''
      description: Create a new transaction
      operationId: postTransactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTransaction'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          description: 'Not connected'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Account not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - auth: []
      tags:
        - transactions
  /accounts/{account_id}/transactions:
    get:
      summary: ''
      description: Get all transactions
      operationId: getAccountTransactions
      parameters:
        - name: account_id
          in: path
          description: ID of the account
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            format: uint64
        - name: limit
          in: query
          description: Number of transactions per page
          required: false
          schema:
            type: integer
            format: uint64
        - name: state
          in: query
          description: Filter by state
          required: false
          schema:
            type: string
            $ref: '#/components/schemas/TransactionState'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  page:
                    type: integer
                    format: uint64
                  limit:
                    type: integer
                    format: uint64
                  max_page:
                    type: integer
                    format: uint64
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
  /accounts/{account_id}/transactions/{transaction_id}:
    get:
      summary: ''
      description: Get transaction
      operationId: getTransactionId
      parameters:
        - name: account_id
          in: path
          description: ID of the account
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: transaction_id
          in: path
          description: ID of the transaction
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Transaction not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
    patch:
      description: Update transaction's state
      operationId: patchTransactionId
      parameters:
        - name: account_id
          in: path
          description: ID of the account
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: transaction_id
          in: path
          description: ID of the transaction
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: state
          in: query
          description: New state of the transaction
          required: true
          schema:
            type: string
            $ref: '#/components/schemas/TransactionState'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Account or transaction not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
    delete:
      summary: ''
      description: Delete transaction
      operationId: markDeleteTransactionId
      parameters:
        - name: account_id
          in: path
          description: ID of the account
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: transaction_id
          in: path
          description: ID of the transaction
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Transaction not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
  /accounts/{account_id}/transactions/{transaction_id}/{item_id}:
    patch:
      description: Update transaction's item state
      operationId: patchTransactionItemId
      parameters:
        - name: account_id
          in: path
          description: ID of the account
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: transaction_id
          in: path
          description: ID of the transaction
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: item_id
          in: path
          description: ID of the item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: state
          in: query
          description: New state of the item
          required: false
          schema:
            $ref: '#/components/schemas/TransactionItemState'
        - name: amount
          in: query
          description: New amount of the item
          required: false
          schema:
            type: integer
            format: uint64
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Account, transaction or item not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - transactions
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 4be0643f-1d98-573b-97cd-ca98a65347dd
    ErrorCodes:
      type: string
      enum:
        - bad_request
        - forbidden
        - account_not_found
        - qr_invalid
        - internal_server_error
        - not_authenticated
        - image_not_found
        - text_not_found
        - category_not_found
        - item_not_found
        - refill_not_found
        - transaction_not_found
      x-enum-varnames:
        - ErrBadRequest
        - ErrForbidden
        - ErrAccountNotFound
        - ErrQRInvalid
        - ErrInternalServerError
        - ErrNotAuthenticated
        - ErrImageNotFound
        - ErrTextNotFound
        - ErrCategoryNotFound
        - ErrItemNotFound
        - ErrRefillNotFound
        - ErrTransactionNotFound
    Messages:
      type: string
      enum:
        - Bad request
        - Not connected
        - Missing fields %v
        - Account already exists
        - Account not found
        - Account is not allowed to use ressource
        - Account cannot use ressource at the time being
        - Invalid QR Code nonce
        - Internal server error
        - You are not authenticated
        - Image does not exists
        - Text does not exists
        - Category does not exists
        - Item does not exists
        - Refill does not exists
        - Transaction does not exists
      x-enum-varnames:
        - MsgBadRequest
        - MsgNotConnected
        - MsgMissingFields
        - MsgAccountAlreadyExists
        - MsgAccountNotFound
        - MsgAccountNotAllowed
        - MsgAccountNotAvailable
        - MsgQRInvalid
        - MsgInternalServerError
        - MsgNotAuthenticated
        - MsgImageNotFound
        - MsgTextNotFound
        - MsgCategoryNotFound
        - MsgItemNotFound
        - MsgRefillNotFound
        - MsgTransactionNotFound
    HTTPError:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/Messages'
        error_code:
          $ref: '#/components/schemas/ErrorCodes'
      required:
        - message
        - error_code
...
    TransactionState:
      type: string
      enum:
        - started
        - taken_care_of
        - finished
        - canceled
      x-enum-varnames:
        - TransactionStarted
        - TransactionTakenCareOf
        - TransactionFinished
        - TransactionCanceled
    TransactionItemState:
      type: string
      enum:
        - started
        - taken_care_of
        - finished
        - canceled
      x-enum-varnames:
        - TransactionItemStarted
        - TransactionItemTakenCareOf
        - TransactionItemFinished
        - TransactionItemCanceled
    TransactionItem:
      type: object
      properties:
        item_id:
          $ref: '#/components/schemas/UUID'
        item_amount:
          type: integer
          format: uint64
        unit_cost:
          type: integer
          format: uint64
        total_cost:
          type: integer
          format: uint64
        state:
          $ref: '#/components/schemas/TransactionItemState'
      required:
        - item_id
        - item_amount
        - unit_cost
        - total_cost
        - state
    Transaction:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        account_id:
          type: string
        total_cost:
          type: integer
          format: uint64
        state:
          $ref: '#/components/schemas/TransactionState'
        deleted_at:
          type: integer
          format: uint64
        deleted_by:
          $ref: '#/components/schemas/UUID'
      required:
        - id
        - items
        - account_id
        - total_cost
        - state
    NewTransactionItem:
      type: object
      properties:
        item_id:
          $ref: '#/components/schemas/UUID'
        amount:
          type: integer
          format: uint64
          
      required:
        - item_id
        - amount
    NewTransaction:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NewTransactionItem'
        card_pin:
          type: string
          description: Pin of the card
      required:
        - items
        - card_pin
...
  securitySchemes:
    auth:
      type: apiKey
      name: BAR_SESS
      in: cookie
    local_token:
      type: apiKey
      name: X-Local-Token
      in: header
    admin_auth:
      type: apiKey
      name: BAR_ADMIN_SESS
      in: cookie